WebApp/requirements.txt
asgiref==3.7.2
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.30.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Django==4.2.6
django-appconf==1.0.5
django-cryptography==1.1
django-environ==0.11.2
idna==3.6
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.2.1
wheel==0.41.2
